add withOverride
